backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
PyPDF2>=3.0.1
python-docx>=1.1.0
faiss-cpu>=1.7.0
sentence-transformers>=3.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.0.0
typing-extensions>=4.0.0
